Rename `rustc_mir_dataflow/src/errors.rs` into `rustc_mir_dataflow/src/diagnostics.rs`
